fix: the service name should be based on the release name
In order to generate the scrape config for prometheus, we need to have a
well-known service name, however there's a limitation on Helm where you can't
access variables from the parent context, so we cannot know what the actual
chart name is.

Prefixing a hard-coded service name with the release name is a middle-ground
compromise which should still allow for multiple kubecost installations if
desired.

Fixes #83