Add a custom 504 page to NGINX config
A common source of confusion when doing support is finding out where 504
errors are coming from. Often this is caused by slowness of a backing
API, but debugging should be substnatially improved with a custom 504
page that will let us discover if our NGINX is the source of a 504 vs.
an external gateway, like a user's load balancer etc.

Tested with a custom "location" to force a 504:
        location /504 {
            return 504;
        }

Observed the new custom_504.html page at /504 when I deployed the change.

The new custom_504.html page is introduced in a separate frontend PR.