Disable ContainerStats pipeline by default
Access to the scale cluster, combined with a last-minute discovery fix
for ContainerStats's multi-cluster support, lead to us discovering that
standard ContainerStats build logic is untenable at scale because it
loads all CPU and RAM data points for all containers over a 24h period.
While we work on improving this, we are disabling it by default.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>