Add schema for status to TurndownSchedule CRD
Fixes a bug where the CR (turndownschedule) status would never be
updated because calls to UpdateStatus were no-ops. The "bad state"
cleanup routine for turndownschedules would then delete the schedule
because it thought it was in a bad state, but in fact we just failed to
update the CR with the state being held by the controller.

With the schema, they are no longer no-ops and turndown completes
successfully.

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>