Reduce Prom retention to 97h for stability
As noted in the new comment, Kubecost has a fairly reliable data store
at this time which we rely upon for long-running metrics storage.
Prometheus has major problems with stability and resource consumption,
particularly at high scale, and we only need it to be queryable long
enough for ETL to build data. Once that data is built, and assuming a
rebuild is not necessary, we no longer need Prometheus to hold that data.
We are wasting a lot of user resources for near-zero gain.

We have performed changes like this (as aggressive as 50h) in user
environments with great success. This is a more conservative 97h to
provide some grace room for agent crashes while still reaping most
of the benefits (from 15d to 4d).

Signed-off-by: Michael Dresser <michaelmdresser@gmail.com>